server/requirements.txt
fastapi
uvicorn
websockets
python-multipart
langchain-google-genai
langchain-core
langgraph
python-dotenv
